pyproject.toml
[tool.poetry]
name = "diplomdf2"
version = "0.1.0"
description = ""
authors = ["Эдуард Кусков <shadowfish@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
psycopg2 = "^2.9.9"
pillow = "^10.4.0"
flake8 = "^7.1.1"
python-dotenv = "^1.0.1"
ipython = "^8.27.0"
pytils = "^0.4.1"
redis = "^5.0.8"
django-phonenumber-field = "^8.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
flake8-docstrings = "^1.7.0"
flake8-builtins = "^2.5.0"
flake8-quotes = "^3.4.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
